optimizing

+++ src/Aufgabe8/Distance.java
@@ -1,5 +1,5 @@
 public class Distance {
-    public Distance(int i, int j, int distance) {
+    public Distance(int i, int j, double distance) {
         if (j < i) {
             this.i = j;
             this.j = i;
@@ -12,5 +12,5 @@ public class Distance {
 
     public final int i;
     public final int j;
-    public final int distance;
+    public final double distance;
 }
